Adding gitignore and root verification on bootstrap process

+++ configure.sh
@@ -2,6 +2,12 @@
 # Author: Rafael Lopes
 
 installdir="/home/pi/SysprePi"
+
+if [ $(whoami) != 'root' ]; then
+  echo "You must be root to do this."
+  exit 1
+fi
+
 if [ -d "$installdir" ]; then
   rm /etc/rc.local
   sudo ln -s $installdir/linux-files/rc.local /etc/rc.local
